+++ ost/tests/client.cpp
@@ -70,11 +70,10 @@ Client::send_allocate(const RawstdUUID& id, uint64_t size, unsigned int width) {
             .object_id = {},
             .chunk_offset = 0,
             .size = size,
-            .reserved1 = 0,
+            .chunk_shift = 0,
             .width = static_cast<uint8_t>(width),
             .reserved2 = 0,
             .reserved3 = 0,
-            .reserved4 = 0,
         },
     };
     std::memcpy(frame.payload.object_id, id.bytes, sizeof(id.bytes));